*** empty log message ***
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk@571 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ jasperreports/src/net/sf/jasperreports/engine/export/JRHtmlExporter.java
@@ -1460,7 +1460,7 @@ public class JRHtmlExporter extends JRAbstractExporter
 		
 				if (image.getHeight() > 0)
 				{
-					double ratio = (double)normalWidth / (double)normalHeight;
+					double ratio = normalWidth / normalHeight;
 					
 					if( ratio > (double)image.getWidth() / (double)image.getHeight() )
 					{
